[NOP] Move search configurator into search detail
